Add global logged in user variable to API (Add Google Auth, access/refresh token functionality, https for dev environment in node)

+++ API/src/controllers/checklistController.ts
@@ -5,10 +5,15 @@ import { Checklist } from "../models/checklist";
 import { AuthRequest } from "../middleware/authorize";
 
 // GET all
-export const getChecklists = async (_req: AuthRequest, res: Response) => {
+export const getChecklists = async (req: AuthRequest, res: Response) => {
     try {
+        const userId = req.user.sub;
+        if (userId === null || userId === undefined) {
+            return res.status(401).json("User not found.");
+        }
+
         const db: Database = await connectDB();
-        const checklists: Checklist[] = await db.all("SELECT * FROM checklist");
+        const checklists: Checklist[] = await db.all("SELECT * FROM checklist WHERE userId = ?", userId);
 
         res.status(200).json(checklists);
     } catch (err) {
@@ -19,6 +24,11 @@ export const getChecklists = async (_req: AuthRequest, res: Response) => {
 // GET by ID
 export const getChecklistById = async (req: AuthRequest, res: Response) => {
     try {
+        const userId = req.user.sub;
+        if (userId === null || userId === undefined) {
+            return res.status(401).json("User not found.");
+        }
+
         const id = Number(req.params.id);
 
         if (isNaN(id)) {
@@ -26,8 +36,16 @@ export const getChecklistById = async (req: AuthRequest, res: Response) => {
         }
 
         const db: Database = await connectDB();
-        const checklist: Checklist | undefined = await db.get("SELECT * FROM checklist WHERE id = ?", id);
-        const checklistItems = await db.all("SELECT * FROM checklistItem WHERE checklistId = ?", id);
+        const checklist: Checklist | undefined = await db.get(
+            "SELECT * FROM checklist WHERE id = ? AND userId = ?",
+            id,
+            userId
+        );
+        const checklistItems = await db.all(
+            "SELECT * FROM checklistItem WHERE checklistId = ? AND userId = ?",
+            id,
+            userId
+        );
 
         if (!checklist) {
             return res.status(404).json({ message: "Item not found" });
@@ -47,6 +65,11 @@ export const getChecklistById = async (req: AuthRequest, res: Response) => {
 // POST create
 export const createChecklist = async (req: AuthRequest, res: Response) => {
     try {
+        const userId = req.user.sub;
+        if (userId === null || userId === undefined) {
+            return res.status(401).json("User not found.");
+        }
+
         const { title } = req.body;
 
         if (typeof title !== "string" || !title.trim()) {
@@ -54,7 +77,7 @@ export const createChecklist = async (req: AuthRequest, res: Response) => {
         }
 
         const db: Database = await connectDB();
-        const result = await db.run("INSERT INTO checklist (title) VALUES (?)", title);
+        const result = await db.run("INSERT INTO checklist (title, userId) VALUES (?, ?)", title, userId);
 
         res.status(201).json({ id: result.lastID, title });
     } catch (err) {
@@ -65,6 +88,11 @@ export const createChecklist = async (req: AuthRequest, res: Response) => {
 // PUT update
 export const updateChecklist = async (req: AuthRequest, res: Response) => {
     try {
+        const userId = req.user.sub;
+        if (userId === null || userId === undefined) {
+            return res.status(401).json("User not found.");
+        }
+
         const id = Number(req.params.id);
         const { title } = req.body;
 
@@ -73,7 +101,11 @@ export const updateChecklist = async (req: AuthRequest, res: Response) => {
         }
 
         const db: Database = await connectDB();
-        const item: Checklist | undefined = await db.get("SELECT * FROM checklist WHERE id = ?", id);
+        const item: Checklist | undefined = await db.get(
+            "SELECT * FROM checklist WHERE id = ? AND userId = ?",
+            id,
+            userId
+        );
 
         if (!item) {
             return res.status(404).json({ message: "Item not found" });
@@ -94,19 +126,28 @@ export const updateChecklist = async (req: AuthRequest, res: Response) => {
 // DELETE
 export const deleteChecklist = async (req: AuthRequest, res: Response) => {
     try {
+        const userId = req.user.sub;
+        if (userId === null || userId === undefined) {
+            return res.status(401).json("User not found.");
+        }
+
         const id = Number(req.params.id);
         if (isNaN(id)) {
             return res.status(400).json({ message: "Invalid ID" });
         }
 
         const db: Database = await connectDB();
-        const item: Checklist | undefined = await db.get("SELECT * FROM checklist WHERE id = ?", id);
+        const item: Checklist | undefined = await db.get(
+            "SELECT * FROM checklist WHERE id = ? AND userId = ?",
+            id,
+            userId
+        );
 
         if (!item) {
             return res.status(404).json({ message: "Item not found" });
         }
 
-        await db.run("DELETE FROM checklist WHERE id = ?", id);
+        await db.run("DELETE FROM checklist WHERE id = ? AND userId = ?", id, userId);
 
         res.status(200).json({ message: "Item successfully removed", id });
     } catch (err) {
